:bug: fix for undo with only two cursors
the first undo is the activation for the fake cursor, moving the cursor count to
1.

the second undo performs the undo the fake cursor did

the third undo deactivates the fake cursor going to the main cursor and bringing
the total count to two.

the fourth undo undoes what the main cursor did.